fix typos in send password reset email
The path to the reset template is:

    app/views/email/reset.txt

So, when the revel app is executed, the current working directory needs
to be correct relative to the hardcoded path to the template.

I'm not sure how to properly use relative paths for this.

The mail server's address and port need to be separated by a colon ':'